Implement scaling and aspect-ratio for the rendered frame
The rendering was previously using a fixed coordinate system,
tied to the hard coded output image size. This works fine when the
output size is left constant, but if we want to change the aspect ratio
or field of view, numbers need to be modified to match the provided
output size.

This change computes the aspect ratio from the pixmap output size,
normalizes the uv coordinates, and uses a w coordinate which provides
control over field of view.